Added missing typecast.

+++ coders/jp2.c
@@ -1037,7 +1037,8 @@ static MagickBooleanType WriteJP2Image(const ImageInfo *image_info,Image *image,
         register int
           *q;
 
-        scale=(double) ((1UL << jp2_image->comps[i].prec)-1)/QuantumRange;
+        scale=(double) (((size_t) 1UL << jp2_image->comps[i].prec)-1)/
+          QuantumRange;
         q=jp2_image->comps[i].data+(y/jp2_image->comps[i].dy*
           image->columns/jp2_image->comps[i].dx+x/jp2_image->comps[i].dx);
         switch (i)
